+++ src/lightwin/beam_calculation/envelope_3d/element_envelope3d_parameters_factory.py
@@ -25,6 +25,7 @@ from lightwin.core.elements.field_maps.field_map_7700 import FieldMap7700
 from lightwin.core.elements.quad import Quad
 from lightwin.core.elements.solenoid import Solenoid
 from lightwin.util.synchronous_phases import PHI_S_MODELS
+from lightwin.util.typing import BeamKwargs
 
 PARAMETERS_3D = {
     Bend: BendEnvelope3DParameters,
@@ -46,7 +47,7 @@ class ElementEnvelope3DParametersFactory(
         method: ENVELOPE3D_METHODS_T,
         n_steps_per_cell: int,
         solver_id: str,
-        beam_kwargs: dict[str, Any],
+        beam_kwargs: BeamKwargs,
         phi_s_definition: PHI_S_MODELS = "historical",
     ) -> None:
         """Prepare import of proper functions."""
@@ -61,7 +62,7 @@ class ElementEnvelope3DParametersFactory(
 
         Parameters
         ----------
-        elt : Element
+        elt :
             Element under study.
 
         Returns