+++ src/lightwin/beam_calculation/simulation_output/factory.py
@@ -9,7 +9,6 @@ import logging
 from abc import ABC, ABCMeta, abstractmethod
 from dataclasses import dataclass
 from functools import partial
-from typing import Any
 
 from lightwin.beam_calculation.simulation_output.simulation_output import (
     SimulationOutput,
@@ -17,6 +16,7 @@ from lightwin.beam_calculation.simulation_output.simulation_output import (
 from lightwin.core.elements.element import ELEMENT_TO_INDEX_T, POS_T, Element
 from lightwin.core.list_of_elements.helper import equivalent_elt
 from lightwin.core.list_of_elements.list_of_elements import ListOfElements
+from lightwin.util.typing import BeamKwargs
 
 
 @dataclass
@@ -26,7 +26,7 @@ class SimulationOutputFactory(ABC):
     _is_3d: bool
     _is_multipart: bool
     _solver_id: str
-    _beam_kwargs: dict[str, Any]
+    _beam_kwargs: BeamKwargs
 
     def __post_init__(self) -> None:
         """Create the factories.
@@ -82,6 +82,7 @@ def _element_to_index(
     elt: Element | str,
     pos: POS_T | None = None,
     return_elt_idx: bool = False,
+    handle_missing_elt: bool = False,
 ) -> int | slice:
     """Convert ``elt`` and ``pos`` into a mesh index.
 
@@ -114,6 +115,8 @@ def _element_to_index(
     return_elt_idx :
         If True, the returned index is the position of the element in
         ``_elts``.
+    handle_missing_elt :
+        Look for an equivalent element when ``elt`` is not in ``_elts``.
 
     Returns
     -------
@@ -122,6 +125,12 @@ def _element_to_index(
 
     """
     if isinstance(elt, str):
+        elt = equivalent_elt(elts=_elts, elt=elt)
+    elif elt not in _elts and handle_missing_elt:
+        logging.debug(
+            f"{elt = } is not in _elts. Trying to take an element in _elts "
+            "with the same name..."
+        )
         elt = equivalent_elt(elts=_elts, elt=elt)
 
     beam_calc_param = elt.beam_calc_param[_solver_id]