+++ src/lightwin/core/beam_parameters/phase_space/phase_space_beam_parameters.py
@@ -9,9 +9,9 @@ For a list of the units associated with every parameter, see
 """
 
 from dataclasses import dataclass
-from typing import Any, Self
+from typing import Self
 
-import numpy as np
+from numpy.typing import NDArray
 
 from lightwin.core.beam_parameters.helper import (
     mismatch_from_arrays,
@@ -21,6 +21,7 @@ from lightwin.core.beam_parameters.helper import (
 from lightwin.core.beam_parameters.phase_space.i_phase_space_beam_parameters import (
     IPhaseSpaceBeamParameters,
 )
+from lightwin.util.typing import BeamKwargs
 
 
 @dataclass
@@ -28,25 +29,25 @@ class PhaseSpaceBeamParameters(IPhaseSpaceBeamParameters):
     """Hold Twiss, emittance, envelopes of a single phase-space."""
 
     # Override some types from mother class
-    eps_no_normalization: np.ndarray
-    eps_normalized: np.ndarray
-    mismatch_factor: np.ndarray | None = None
+    eps_no_normalization: NDArray
+    eps_normalized: NDArray
+    mismatch_factor: NDArray | None = None
 
     # Already with proper type in mother class:
-    # envelopes: np.ndarray | None = None
-    # twiss: np.ndarray | None = None
-    # tm_cumul: np.ndarray | None = None
-    # sigma: np.ndarray | None = None
+    # envelopes: NDArray | None = None
+    # twiss: NDArray | None = None
+    # tm_cumul: NDArray | None = None
+    # sigma: NDArray | None = None
 
     @classmethod
     def from_cumulated_transfer_matrices(
         cls,
         phase_space_name: str,
-        sigma_in: np.ndarray,
-        tm_cumul: np.ndarray,
-        gamma_kin: np.ndarray,
-        beta_kin: np.ndarray,
-        beam_kwargs: dict[str, Any],
+        sigma_in: NDArray,
+        tm_cumul: NDArray,
+        gamma_kin: NDArray,
+        beta_kin: NDArray,
+        beam_kwargs: BeamKwargs,
     ) -> Self:
         r"""Compute :math:`\sigma` matrix, and everything from it."""
         sigma = sigma_from_transfer_matrices(sigma_in, tm_cumul)
@@ -64,11 +65,11 @@ class PhaseSpaceBeamParameters(IPhaseSpaceBeamParameters):
     def from_sigma(
         cls,
         phase_space_name: str,
-        sigma: np.ndarray,
-        gamma_kin: np.ndarray,
-        beta_kin: np.ndarray,
-        beam_kwargs: dict[str, Any],
-        **kwargs: np.ndarray,  # tm_cumul
+        sigma: NDArray,
+        gamma_kin: NDArray,
+        beta_kin: NDArray,
+        beam_kwargs: BeamKwargs,
+        **kwargs: NDArray,  # tm_cumul
     ) -> Self:
         """Compute Twiss, eps, envelopes just from sigma matrix."""
         return super().from_sigma(
@@ -85,10 +86,10 @@ class PhaseSpaceBeamParameters(IPhaseSpaceBeamParameters):
         cls,
         other_phase_space: Self,
         phase_space_name: str,
-        gamma_kin: np.ndarray,
-        beta_kin: np.ndarray,
-        beam_kwargs: dict[str, Any],
-        **kwargs: np.ndarray,  # sigma, tm_cumul
+        gamma_kin: NDArray,
+        beta_kin: NDArray,
+        beam_kwargs: BeamKwargs,
+        **kwargs: NDArray,  # sigma, tm_cumul
     ) -> Self:
         """Fully initialize from another phase space."""
         return super().from_other_phase_space(
@@ -138,82 +139,82 @@ class PhaseSpaceBeamParameters(IPhaseSpaceBeamParameters):
         return phase_space
 
     @property
-    def alpha(self) -> np.ndarray | None:
+    def alpha(self) -> NDArray | None:
         """Get first column of ``self.twiss``."""
         if self.twiss is None:
             return None
         return self.twiss[:, 0]
 
     @alpha.setter
-    def alpha(self, value: np.ndarray) -> None:
+    def alpha(self, value: NDArray) -> None:
         """Set first column of ``self.twiss``."""
         assert self.twiss is not None
         self.twiss[:, 0] = value
 
     @property
-    def beta(self) -> np.ndarray | None:
+    def beta(self) -> NDArray | None:
         """Get second column of ``self.twiss``."""
         if self.twiss is None:
             return None
         return self.twiss[:, 1]
 
     @beta.setter
-    def beta(self, value: np.ndarray) -> None:
+    def beta(self, value: NDArray) -> None:
         """Set second column of ``self.twiss``."""
         assert self.twiss is not None
         self.twiss[:, 1] = value
 
     @property
-    def gamma(self) -> np.ndarray | None:
+    def gamma(self) -> NDArray | None:
         """Get third column of ``self.twiss``."""
         if self.twiss is None:
             return None
         return self.twiss[:, 2]
 
     @gamma.setter
-    def gamma(self, value: np.ndarray) -> None:
+    def gamma(self, value: NDArray) -> None:
         """Set third column of ``self.twiss``."""
         assert self.twiss is not None
         self.twiss[:, 2] = value
 
     @property
-    def envelope_pos(self) -> np.ndarray | None:
+    def envelope_pos(self) -> NDArray | None:
         """Get first column of ``self.envelopes``."""
         if self.envelopes is None:
             return None
         return self.envelopes[:, 0]
 
     @envelope_pos.setter
-    def envelope_pos(self, value: np.ndarray) -> None:
+    def envelope_pos(self, value: NDArray) -> None:
         """Set first column of ``self.envelopes``."""
         assert self.envelopes is not None
         self.envelopes[:, 0] = value
 
     @property
-    def envelope_energy(self) -> np.ndarray | None:
+    def envelope_energy(self) -> NDArray | None:
         """Get second column of ``self.envelopes``."""
         if self.envelopes is None:
             return None
         return self.envelopes[:, 1]
 
     @envelope_energy.setter
-    def envelope_energy(self, value: np.ndarray) -> None:
+    def envelope_energy(self, value: NDArray) -> None:
         """Set second column of ``self.envelopes``."""
         assert self.envelopes is not None
         self.envelopes[:, 1] = value
 
     @property
-    def eps(self) -> np.ndarray:
+    def eps(self) -> NDArray:
         """Return the normalized emittance."""
         return self.eps_normalized
 
     @property
-    def non_norm_eps(self) -> np.ndarray:
+    def non_norm_eps(self) -> NDArray:
         """Return the non-normalized emittance."""
         return self.eps_no_normalization
 
     @property
-    def sigma_in(self) -> np.ndarray:
+    def sigma_in(self) -> NDArray:
         r"""Return the first :math:`\sigma` beam matrix."""
         assert self.sigma is not None
         return self.sigma[0]
@@ -221,8 +222,8 @@ class PhaseSpaceBeamParameters(IPhaseSpaceBeamParameters):
     def set_mismatch(
         self,
         reference_phase_space: Self,
-        reference_z_abs: np.ndarray,
-        z_abs: np.ndarray,
+        reference_z_abs: NDArray,
+        z_abs: NDArray,
         raise_missing_twiss_error: bool = True,
         **mismatch_kw: bool,
     ) -> None: